requirements.txt
flask
mysql-connector-python
reportlab
gunicorn
